1. Design the game window
Firstly create a java project name Breakout-Ball into your ide. Inside that project folder, create a package and class Main.java with the main method. In this main method, we will create and design a game window. Here, I have also created an object of gameplay to run the game which will be discussed further in detail. So, don't worry.