Add files via upload
When the game is started, in order to move the ball, we have to detect if it’s touching left, right, or top of the panel and change its direction accordingly. Then, we have to find the collision between the ball and the paddle, and the ball and the map. For detecting collisions between two objects, we need a rectangle, so we have to create a rectangle around the ball as it is oval.